Refactored the class HomePage and print out We are in the home page
